+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Add CITATION.CFF and Script to Bump Version [#526](https://github.com/gridstatus/gridstatus/pull/526)
 * Add read_csv Kwargs for ERCOT read_doc [#528](https://github.com/gridstatus/gridstatus/pull/528)
 * upgrade lxml to 5.3.0 [#530](https://github.com/gridstatus/gridstatus/pull/530)
+* Query current day data from real-time hourly market in ISO-NE [#532](https://github.com/gridstatus/gridstatus/pull/532)
+
 
 ## v0.29.0 - January 15, 2025